change the event handler to the event processor, change the builder pattern to the constructor
